+++ HISTORY.md
@@ -1,5 +1,10 @@
 # History
 
+## v1.1.1 2018 September 8
+- Added support for logging to help debug resolution errors
+- Added a fallback in case type checker provided a false negative on class testing
+- Updated `typechecker` dependency from v4.5.0 to v4.6.0
+
 ## v1.1.0 2018 September 8
 - Removed `isPluginClass` for `require('typechecker').isClass` as strict checking on this resulted in too many false negatives due to different baseplugin versions or locations being used between plugins and apps